requirements/base.txt
Flask==2.0.0
kafka-python==2.0.2
fake==0.8
python-dotenv==0.14.0
